Small fixes and absl status.

+++ src/transaction/TransactionId.hpp
@@ -9,6 +9,6 @@ class TransactionId {
     long long get_id() const { return my_id_; }
 
    private:
-    static inline std::atomic<int> counter{0};
+    static inline std::atomic<long long> counter{0};
     long long my_id_;
 };
